Issue #2044 Move FS sync functionality to DtsSynchronizationService
